+++ designer/src/control-bar/control-bar-view.spec.ts
@@ -6,7 +6,7 @@ describe('ControlBarView', () => {
 		const parent = Dom.element('div');
 		const i18n = (_: string, defaultValue: string) => defaultValue;
 
-		const component = ControlBarView.create(parent, true, i18n);
+		const component = ControlBarView.create(parent, true, null, i18n);
 
 		expect(component).toBeDefined();
 		expect(parent.children.length).not.toEqual(0);