+++ designer/src/control-bar/control-bar.ts
@@ -7,19 +7,25 @@ import { ViewportApi } from '../api';
 export class ControlBar implements UiComponent {
 	public static create(parent: HTMLElement, api: DesignerApi): UiComponent {
 		const isUndoRedoSupported = api.controlBar.isUndoRedoSupported();
-		const view = ControlBarView.create(parent, isUndoRedoSupported, api.i18n);
+		const customButtons = api.controlBar.getButtons();
+
+		const view = ControlBarView.create(parent, isUndoRedoSupported, customButtons, api.i18n);
 		const bar = new ControlBar(view, api.controlBar, api.viewport, isUndoRedoSupported);
 
-		view.bindResetButtonClick(() => bar.onResetButtonClicked());
-		view.bindZoomInButtonClick(() => bar.onZoomInButtonClicked());
-		view.bindZoomOutButtonClick(() => bar.onZoomOutButtonClicked());
-		view.bindDisableDragButtonClick(() => bar.onMoveButtonClicked());
-		view.bindDeleteButtonClick(() => bar.onDeleteButtonClicked());
-		api.controlBar.onStateChanged.subscribe(() => bar.refreshButtons());
+		view.bindResetButtonClick(bar.onResetButtonClicked);
+		view.bindZoomInButtonClick(bar.onZoomInButtonClicked);
+		view.bindZoomOutButtonClick(bar.onZoomOutButtonClicked);
+		view.bindDisableDragButtonClick(bar.onMoveButtonClicked);
+		view.bindDeleteButtonClick(bar.onDeleteButtonClicked);
+		api.controlBar.onStateChanged.subscribe(bar.refreshButtons);
 
 		if (isUndoRedoSupported) {
-			view.bindUndoButtonClick(() => bar.onUndoButtonClicked());
-			view.bindRedoButtonClick(() => bar.onRedoButtonClicked());
+			view.bindUndoButtonClick(bar.onUndoButtonClicked);
+			view.bindRedoButtonClick(bar.onRedoButtonClicked);
+		}
+
+		if (customButtons) {
+			view.bindCustomButtonClick(bar.onCustomButtonClicked);
 		}
 
 		bar.refreshButtons();
@@ -41,41 +47,45 @@ export class ControlBar implements UiComponent {
 		//
 	}
 
-	private onResetButtonClicked() {
+	private readonly onResetButtonClicked = () => {
 		this.viewportApi.resetViewport();
-	}
+	};
 
-	private onZoomInButtonClicked() {
+	private readonly onZoomInButtonClicked = () => {
 		this.viewportApi.zoom(true);
-	}
+	};
 
-	private onZoomOutButtonClicked() {
+	private readonly onZoomOutButtonClicked = () => {
 		this.viewportApi.zoom(false);
-	}
+	};
 
-	private onMoveButtonClicked() {
+	private readonly onMoveButtonClicked = () => {
 		this.controlBarApi.toggleIsDragDisabled();
-	}
+	};
 
-	private onUndoButtonClicked() {
+	private readonly onUndoButtonClicked = () => {
 		this.controlBarApi.tryUndo();
-	}
+	};
 
-	private onRedoButtonClicked() {
+	private readonly onRedoButtonClicked = () => {
 		this.controlBarApi.tryRedo();
-	}
+	};
 
-	private onDeleteButtonClicked() {
+	private readonly onDeleteButtonClicked = () => {
 		this.controlBarApi.tryDelete();
-	}
+	};
 
-	private refreshButtons() {
+	private readonly refreshButtons = () => {
 		this.refreshDeleteButtonVisibility();
 		this.refreshIsDragDisabled();
 		if (this.isUndoRedoSupported) {
 			this.refreshUndoRedoAvailability();
 		}
-	}
+	};
+
+	private readonly onCustomButtonClicked = (id: string) => {
+		this.controlBarApi.triggerButtonClick(id);
+	};
 
 	//